very small PR which moves the TeamMemberSection to its own file for upcoming work (and to decouple the TeamPage)
