+++ packages/app/src/components/TeamSettings/TeamMembersSection.tsx
@@ -1,0 +1,442 @@
+import { useState } from 'react';
+import { HTTPError } from 'ky';
+import CopyToClipboard from 'react-copy-to-clipboard';
+import {
+  Badge,
+  Box,
+  Button,
+  Card,
+  Divider,
+  Group,
+  Modal,
+  Stack,
+  Table,
+  Text,
+  TextInput,
+} from '@mantine/core';
+import { notifications } from '@mantine/notifications';
+
+import api from '@/api';
+
+export default function TeamMembersSection() {
+  const hasAdminAccess = true;
+
+  const { data: team } = api.useTeam();
+  const {
+    data: members,
+    isLoading: isLoadingMembers,
+    refetch: refetchMembers,
+  } = api.useTeamMembers();
+
+  const {
+    data: invitations,
+    isLoading: isLoadingInvitations,
+    refetch: refetchInvitations,
+  } = api.useTeamInvitations();
+
+  const onSubmitTeamInviteForm = ({ email }: { email: string }) => {
+    sendTeamInviteAction(email);
+    setTeamInviteModalShow(false);
+  };
+
+  const [
+    deleteTeamMemberConfirmationModalData,
+    setDeleteTeamMemberConfirmationModalData,
+  ] = useState<{
+    mode: 'team' | 'teamInvite' | null;
+    id: string | null;
+    email: string | null;
+  }>({
+    mode: null,
+    id: null,
+    email: null,
+  });
+  const [teamInviteModalShow, setTeamInviteModalShow] = useState(false);
+
+  const saveTeamInvitation = api.useSaveTeamInvitation();
+  const deleteTeamMember = api.useDeleteTeamMember();
+  const deleteTeamInvitation = api.useDeleteTeamInvitation();
+
+  const sendTeamInviteAction = (email: string) => {
+    if (email) {
+      saveTeamInvitation.mutate(
+        { email },
+        {
+          onSuccess: () => {
+            notifications.show({
+              color: 'green',
+              message:
+                'Click "Copy URL" and share the URL with your team member',
+            });
+            refetchInvitations();
+          },
+          onError: e => {
+            if (e instanceof HTTPError) {
+              e.response
+                .json()
+                .then(res => {
+                  notifications.show({
+                    color: 'red',
+                    message: res.message,
+                    autoClose: 5000,
+                  });
+                })
+                .catch(() => {
+                  notifications.show({
+                    color: 'red',
+                    message:
+                      'Something went wrong. Please contact HyperDX team.',
+
+                    autoClose: 5000,
+                  });
+                });
+            } else {
+              notifications.show({
+                color: 'red',
+                message: 'Something went wrong. Please contact HyperDX team.',
+                autoClose: 5000,
+              });
+            }
+          },
+        },
+      );
+    }
+  };
+
+  const onConfirmDeleteTeamMember = (id: string) => {
+    if (deleteTeamMemberConfirmationModalData.mode === 'team') {
+      deleteTeamMemberAction(id);
+    } else if (deleteTeamMemberConfirmationModalData.mode === 'teamInvite') {
+      deleteTeamInviteAction(id);
+    }
+    setDeleteTeamMemberConfirmationModalData({
+      mode: null,
+      id: null,
+      email: null,
+    });
+  };
+
+  const deleteTeamInviteAction = (id: string) => {
+    if (id) {
+      deleteTeamInvitation.mutate(
+        { id: encodeURIComponent(id) },
+        {
+          onSuccess: () => {
+            notifications.show({
+              color: 'green',
+              message: 'Deleted team invite',
+            });
+            refetchInvitations();
+          },
+          onError: e => {
+            if (e instanceof HTTPError) {
+              e.response
+                .json()
+                .then(res => {
+                  notifications.show({
+                    color: 'red',
+                    message: res.message,
+                    autoClose: 5000,
+                  });
+                })
+                .catch(() => {
+                  notifications.show({
+                    color: 'red',
+                    message:
+                      'Something went wrong. Please contact HyperDX team.',
+
+                    autoClose: 5000,
+                  });
+                });
+            } else {
+              notifications.show({
+                color: 'red',
+                message: 'Something went wrong. Please contact HyperDX team.',
+                autoClose: 5000,
+              });
+            }
+          },
+        },
+      );
+    }
+  };
+  const deleteTeamMemberAction = (id: string) => {
+    if (id) {
+      deleteTeamMember.mutate(
+        { userId: encodeURIComponent(id) },
+        {
+          onSuccess: () => {
+            notifications.show({
+              color: 'green',
+              message: 'Deleted team member',
+            });
+            refetchMembers();
+          },
+          onError: e => {
+            if (e instanceof HTTPError) {
+              e.response
+                .json()
+                .then(res => {
+                  notifications.show({
+                    color: 'red',
+                    message: res.message,
+                    autoClose: 5000,
+                  });
+                })
+                .catch(() => {
+                  notifications.show({
+                    color: 'red',
+                    message:
+                      'Something went wrong. Please contact HyperDX team.',
+                    autoClose: 5000,
+                  });
+                });
+            } else {
+              notifications.show({
+                color: 'red',
+                message: 'Something went wrong. Please contact HyperDX team.',
+                autoClose: 5000,
+              });
+            }
+          },
+        },
+      );
+    }
+  };
+
+  return (
+    <Box id="team_members">
+      <Text size="md">Team</Text>
+      <Divider my="md" />
+
+      <Card>
+        <Card.Section withBorder py="sm" px="lg">
+          <Group align="center" justify="space-between">
+            <div className="fs-7">Team Members</div>
+            <Button
+              variant="light"
+              leftSection={<i className="bi bi-person-plus-fill" />}
+              onClick={() => setTeamInviteModalShow(true)}
+            >
+              Invite Team Member
+            </Button>
+          </Group>
+        </Card.Section>
+        <Card.Section>
+          <Table horizontalSpacing="lg" verticalSpacing="xs">
+            <Table.Tbody>
+              {!isLoadingMembers &&
+                Array.isArray(members?.data) &&
+                members?.data.map((member: any) => (
+                  <Table.Tr key={member.email}>
+                    <Table.Td>
+                      <div>
+                        {member.isCurrentUser && (
+                          <Badge variant="light" mr="xs" tt="none">
+                            You
+                          </Badge>
+                        )}
+                        <span className="text-white fw-bold fs-7">
+                          {member.name}
+                        </span>
+                      </div>
+                      <Group mt={4} fz="xs">
+                        <div>{member.email}</div>
+                        {member.hasPasswordAuth && (
+                          <div>
+                            <i className="bi bi-lock-fill" /> Password Auth
+                          </div>
+                        )}
+                      </Group>
+                    </Table.Td>
+                    <Table.Td>
+                      {team.shouldEnforceRBAC && !member.groupName && (
+                        <Badge
+                          variant="light"
+                          color="red"
+                          fw="normal"
+                          tt="none"
+                        >
+                          Not Assigned to Group
+                        </Badge>
+                      )}
+                      {member.groupName && (
+                        <Badge
+                          variant="light"
+                          color="green"
+                          fw="normal"
+                          tt="none"
+                        >
+                          {member.groupName}
+                        </Badge>
+                      )}
+                    </Table.Td>
+                    <Table.Td style={{ textAlign: 'right' }}>
+                      {!member.isCurrentUser && hasAdminAccess && (
+                        <Group justify="flex-end" gap="8">
+                          <Button
+                            size="compact-sm"
+                            variant="light"
+                            color="red"
+                            onClick={() =>
+                              setDeleteTeamMemberConfirmationModalData({
+                                mode: 'team',
+                                id: member._id,
+                                email: member.email,
+                              })
+                            }
+                          >
+                            Remove
+                          </Button>
+                        </Group>
+                      )}
+                    </Table.Td>
+                  </Table.Tr>
+                ))}
+              {!isLoadingInvitations &&
+                Array.isArray(invitations.data) &&
+                invitations.data.map((invitation: any) => (
+                  <Table.Tr key={invitation.email} className="mt-2">
+                    <Table.Td>
+                      <span className="text-white fw-bold fs-7">
+                        {invitation.email}
+                      </span>
+                    </Table.Td>
+                    <Table.Td>
+                      <Badge variant="dot" color="gray" fw="normal" tt="none">
+                        Pending Invite
+                      </Badge>
+                      <CopyToClipboard text={invitation.url}>
+                        <Button size="compact-xs" variant="default" ml="xs">
+                          📋 Copy URL
+                        </Button>
+                      </CopyToClipboard>
+                    </Table.Td>
+                    <Table.Td style={{ textAlign: 'right' }}>
+                      {hasAdminAccess && (
+                        <Group justify="flex-end" gap="8">
+                          <Button
+                            size="compact-sm"
+                            variant="light"
+                            color="red"
+                            onClick={() =>
+                              setDeleteTeamMemberConfirmationModalData({
+                                mode: 'teamInvite',
+                                id: invitation._id,
+                                email: invitation.email,
+                              })
+                            }
+                          >
+                            Delete
+                          </Button>
+                        </Group>
+                      )}
+                    </Table.Td>
+                  </Table.Tr>
+                ))}
+            </Table.Tbody>
+          </Table>
+        </Card.Section>
+      </Card>
+
+      <Modal
+        centered
+        onClose={() => setTeamInviteModalShow(false)}
+        opened={teamInviteModalShow}
+        title="Invite Team Member"
+      >
+        <InviteTeamMemberForm
+          onSubmit={onSubmitTeamInviteForm}
+          isSubmitting={saveTeamInvitation.isPending}
+        />
+      </Modal>
+
+      <Modal
+        centered
+        onClose={() =>
+          setDeleteTeamMemberConfirmationModalData({
+            mode: null,
+            id: null,
+            email: null,
+          })
+        }
+        opened={deleteTeamMemberConfirmationModalData.id != null}
+        size="lg"
+        title="Delete Team Member"
+      >
+        <Stack>
+          <Text>
+            Deleting this team member (
+            {deleteTeamMemberConfirmationModalData.email}) will revoke their
+            access to the team&apos;s resources and services. This action is not
+            reversible.
+          </Text>
+          <Group justify="flex-end" gap="xs">
+            <Button
+              variant="default"
+              onClick={() =>
+                setDeleteTeamMemberConfirmationModalData({
+                  mode: null,
+                  id: null,
+                  email: null,
+                })
+              }
+            >
+              Cancel
+            </Button>
+            <Button
+              variant="outline"
+              color="red"
+              onClick={() =>
+                deleteTeamMemberConfirmationModalData.id &&
+                onConfirmDeleteTeamMember(
+                  deleteTeamMemberConfirmationModalData.id,
+                )
+              }
+            >
+              Confirm
+            </Button>
+          </Group>
+        </Stack>
+      </Modal>
+    </Box>
+  );
+}
+
+function InviteTeamMemberForm({
+  isSubmitting,
+  onSubmit,
+}: {
+  isSubmitting?: boolean;
+  onSubmit: (arg0: { email: string }) => void;
+}) {
+  const [email, setEmail] = useState<string>('');
+
+  return (
+    <form
+      onSubmit={e => {
+        onSubmit({ email });
+        e.preventDefault();
+      }}
+    >
+      <Stack>
+        <TextInput
+          label="Email"
+          name="email"
+          type="email"
+          value={email}
+          onChange={e => setEmail(e.target.value)}
+          required
+          placeholder="you@company.com"
+          withAsterisk={false}
+        />
+        <div className="fs-8">
+          The invite link will automatically expire after 30 days.
+        </div>
+        <Button variant="light" type="submit" disabled={!email || isSubmitting}>
+          Send Invite
+        </Button>
+      </Stack>
+    </form>
+  );
+}
